When sending CelesteNetClientOptions as header strings during the handshake, enum types would ToString to the names of the values rather than the underlying numeric type.
This changes it so that the value gets sent as the underlying type and then the server parses it by its TypeCode as normal.

+++ CelesteNet.Client/CelesteNetClient.cs
@@ -54,7 +54,7 @@ namespace Celeste.Mod.CelesteNet.Client {
             Options.ClientID = Settings.ClientID;
             Options.InstanceID = Settings.InstanceID;
 
-            Options.SupportedClientFeatures |= (ulong) CelesteNetSupportedClientFeatures.LocateCommand;
+            Options.SupportedClientFeatures |= CelesteNetSupportedClientFeatures.LocateCommand;
 
             bool isServerLocalhost = Settings.Host == "localhost" || Settings.Host == "127.0.0.1";
